Fix incorrect 'required' property on Resource in OpenAPI definition
We were incorrectly setting `name` instead of `uri` in the 'required' properties for `Resource`. This was causing issues as `uri` is actually required, and not setting it causes issue further down the line.

This has been fixed and the client model has been regenerated. Additionally, `displayName` and `uri` order have been swapped in the OpenAPI definition so that the parameter order does not change in the C# model, which may have caused issues for downstream consumers, since both parameters are strings and the change may not have been noticed.

Next version has been updated to 2.0 to reflect the breaking change in the API.