Separate Create and Update of ClaimPermissions
The creation endpoint was checking for the existence of a claim permission with the same id as the new one before proceeding, as part of its overall prerequisite checking. But as Matthew pointed out it is more robust to perform a "does this thing I'm about to create already exist?" type checks atomically at the point of creation with a conditional upload. (I had missed this because I all the other prereqs can't be handled this way, and I was thinking of this as just another prereq. Also, I'd argue that the scenarios in which this extra robustness makes a different are ones we are incredibly unlikely to run into: the only reason I added this check was to catch bone-headed-type errors where the claims initialization code is just doing the wrong thing; claims setup tends to be an infrequent thing, so concurrency issues like this will rarely arise. However, I might be wrong about that, and anyway, it's better to do things right.)