Update Microsoft.Extensions.* and enable lock files
While our libraries generally want to use floating references for Microsoft.Extensions.* components (to avoid needing to do a brand new release of every single component every time Microsoft updates .NET) we need to be deliberate about taking those updates at the top of the food chain. Otherwise, our builds are not repeatable - the source which this week picked up 3.1.19 of all these components might next week pick up 3.1.20. So projects that produce (or test) runnable artifacts such as Azure Functions projects or tests need to use lock files so that when changes of this kind percolate through the dependency tree, the point at which we chose to take them is baked into the source tree.